Made some major changes. App is now more stable and performance is enhanced

+++ app/src/main/java/com/simplicity/anuj/myday/FAB.java
@@ -6,7 +6,6 @@ import android.support.design.widget.AppBarLayout;
 import android.support.design.widget.CoordinatorLayout;
 import android.support.design.widget.FloatingActionButton;
 import android.util.AttributeSet;
-import android.util.Log;
 import android.view.View;
 import android.view.animation.LinearInterpolator;
 
@@ -36,7 +35,7 @@ public class FAB extends CoordinatorLayout.Behavior<FloatingActionButton> {
             CoordinatorLayout.LayoutParams lp = (CoordinatorLayout.LayoutParams) fab.getLayoutParams();
             int fabBottomMargin = lp.bottomMargin;
             int distanceToScroll = fab.getHeight() + fabBottomMargin;
-            float ratio = (float) dependency.getY() / (float) toolbarHeight;
+            float ratio = dependency.getY() / (float) toolbarHeight;
             fab.animate().translationY(-distanceToScroll * ratio).setInterpolator(new LinearInterpolator()).start();
         }
         return true;
